refactor: compile time flags, modularity
